debugging the building process
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@2318 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/helpers/RandomDistributor.h
@@ -17,7 +17,7 @@ public:
 
     T get() const {
         SUMOReal prob = (SUMOReal) (((SUMOReal) rand()/(SUMOReal) RAND_MAX) * myProb);
-        for(int i=0; i<myVals.size(); i++) {
+        for(int i=0; i<(int) myVals.size(); i++) {
             if(prob<myProbs[i]) {
                 return myVals[i];
             }
